Move peer assessment private methods into models. Move errors for assessment app into a module.
